Time: 8 ms (66.46%), Space: 35 MB (78.75%) - LeetHub

+++ 2090-number-of-ways-to-arrive-at-destination/2090-number-of-ways-to-arrive-at-destination.cpp
@@ -24,7 +24,7 @@ public:
 
         while(!pq.empty())
         {
-            int currTime = pq.top().first;
+            long long currTime = pq.top().first;
             int currNode = pq.top().second;
 
             pq.pop();
